login refactor
