--- src/main/java/com/grungapda/backend/user/command/application/dto/login/LoginResponse.java
@@ -1,22 +1,0 @@
-package com.grungapda.backend.user.command.application.dto.login;
-
-import com.grungapda.backend.user.command.application.dto.auth.AuthResponse;
-import lombok.*;
-
-import java.util.List;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class LoginResponse {
-
-    private Long userNo;
-    private String userNickname;
-    private List<AuthResponse> authority;
-
-    @Builder
-    public LoginResponse(Long userNo, String userNickname, List<AuthResponse> authority) {
-        this.userNo = userNo;
-        this.userNickname = userNickname;
-        this.authority = authority;
-    }
-}
