--- src/main/java/com/grungapda/backend/user/command/domain/repository/AuthorityRepository.java
@@ -1,17 +1,0 @@
-package com.grungapda.backend.user.command.domain.repository;
-
-import com.grungapda.backend.user.command.application.dto.login.LogoutResponse;
-import com.grungapda.backend.user.command.domain.aggregate.entity.Authority;
-import com.grungapda.backend.user.command.domain.aggregate.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface AuthorityRepository extends JpaRepository<Authority, Long> {
-    Optional<Authority> findByUser(User user);
-
-    Optional<Authority> findByRefreshToken(String refreshToken);
-
-}
