+++ src/main/java/com/grungapda/backend/userListener/command/application/dto/login/LoginRequest.java
@@ -1,0 +1,10 @@
+package com.grungapda.backend.userListener.command.application.dto.login;
+
+import lombok.Getter;
+
+@Getter
+public class LoginRequest {
+
+    private String userEmail;
+    private String userPwd;
+}
